requirements.txt
Flask==2.2.2
matplotlib==3.7.1
numpy==1.23.5
pandas==1.4.4
Pillow==9.4.0
Werkzeug==2.2.2
flask_session==0.5.0